+++ src/common/utils/index.ts
@@ -1,10 +1,20 @@
-import * as child from 'child_process';
-import * as csvStringify from 'csv-stringify/lib/sync';
-import * as fs from 'fs-extra';
-import * as path from 'path';
-import * as util from 'util';
-import * as xml2js from 'xml2js';
+import * as child from "child_process";
+import * as csvStringify from "csv-stringify/lib/sync";
+import * as fs from "fs-extra";
+import * as path from "path";
+import * as util from "util";
+import * as xml2js from "xml2js";
 const exec = util.promisify(child.exec);
+import simpleGit, { SimpleGit } from "simple-git";
+
+let git: SimpleGit = null;
+export const isGit = child.exec(
+  "git rev-parse --is-inside-work-tree 2>/dev/null",
+  { encoding: "utf8" }
+);
+if (isGit) {
+  git = simpleGit();
+}
 
 let pluginsStdout = null;
 
@@ -14,7 +24,7 @@ export async function checkSfdxPlugin(
 ): Promise<{ installed: boolean; message: string }> {
   let installed = false;
   if (pluginsStdout == null) {
-    const pluginsRes = await exec('sfdx plugins');
+    const pluginsRes = await exec("sfdx plugins");
     pluginsStdout = pluginsRes.stdout;
   }
   if (!pluginsStdout.includes(pluginName)) {
@@ -29,19 +39,35 @@ export async function checkSfdxPlugin(
   };
 }
 
+// Get local git branch name
+export async function getCurrentGitBranch(options: any = { formatted: false }) {
+  const gitBranch = (await git.branchLocal()).current;
+  if (options.formatted === true) {
+    return gitBranch.replace("/", "-");
+  }
+  return gitBranch;
+}
+
 // Execute command and parse result as json
-export async function execJson(command: string, commandThis: any): Promise<any> {
+export async function execJson(
+  command: string,
+  commandThis: any
+): Promise<any> {
   commandThis.ux.log(`[sfdx-hardis][command] ${command}`);
-  if (!command.includes('--json')) {
-    command += ' --json';
+  if (!command.includes("--json")) {
+    command += " --json";
   }
   const commandResult = await exec(command);
   try {
     return JSON.parse(commandResult.stdout);
   } catch (e) {
-    return { status: 1, errorMessage: `[sfdx-hardis][ERROR] Error parsing JSON in command result ${JSON.stringify(commandResult)}` };
+    return {
+      status: 1,
+      errorMessage: `[sfdx-hardis][ERROR] Error parsing JSON in command result ${JSON.stringify(
+        commandResult
+      )}`
+    };
   }
-
 }
 
 // Filter package XML
@@ -64,22 +90,35 @@ export async function filterPackageXml(
   if ((options.removeNamespaces || []).length > 0) {
     manifest.Package.types = manifest.Package.types.map((type: any) => {
       type.members = type.members.filter((member: string) => {
-        return options.removeNamespaces.filter((ns: string) => member.startsWith(ns)).length === 0;
+        return (
+          options.removeNamespaces.filter((ns: string) => member.startsWith(ns))
+            .length === 0
+        );
       });
       return type;
     });
   }
   // Remove from other packageXml file
   if (options.removeFromPackageXmlFile) {
-    const destructiveFileContent = await fs.readFile(options.removeFromPackageXmlFile);
-    const destructiveManifest = await xml2js.parseStringPromise(destructiveFileContent);
+    const destructiveFileContent = await fs.readFile(
+      options.removeFromPackageXmlFile
+    );
+    const destructiveManifest = await xml2js.parseStringPromise(
+      destructiveFileContent
+    );
     manifest.Package.types = manifest.Package.types.map((type: any) => {
-      const destructiveTypes = destructiveManifest.Package.types.filter((destructiveType: any) => {
-        return destructiveType.name[0] === type.name[0];
-      });
+      const destructiveTypes = destructiveManifest.Package.types.filter(
+        (destructiveType: any) => {
+          return destructiveType.name[0] === type.name[0];
+        }
+      );
       if (destructiveTypes.length > 0) {
         type.members = type.members.filter((member: string) => {
-          return destructiveTypes[0].members.filter((destructiveMember: string) => destructiveMember === member).length === 0;
+          return (
+            destructiveTypes[0].members.filter(
+              (destructiveMember: string) => destructiveMember === member
+            ).length === 0
+          );
         });
       }
       return type;
@@ -88,13 +127,13 @@ export async function filterPackageXml(
   // Remove standard objects
   if (options.removeStandard) {
     manifest.Package.types = manifest.Package.types.map((type: any) => {
-      if (['CustomObject'].includes(type.name[0])) {
+      if (["CustomObject"].includes(type.name[0])) {
         type.members = type.members.filter((member: string) => {
-          return member.endsWith('__c');
+          return member.endsWith("__c");
         });
       }
       type.members = type.members.filter((member: string) => {
-        return !member.startsWith('standard__');
+        return !member.startsWith("standard__");
       });
       return type;
     });
@@ -104,8 +143,10 @@ export async function filterPackageXml(
     manifest.Package.version[0] = options.updateApiVersion;
   }
   // Remove metadata types (named, and empty ones)
-  manifest.Package.types = manifest.Package.types.filter((type: any) =>
-    !(options.removeMetadatas || []).includes(type.name[0]) && (type?.members?.length || 0) > 0
+  manifest.Package.types = manifest.Package.types.filter(
+    (type: any) =>
+      !(options.removeMetadatas || []).includes(type.name[0]) &&
+      (type?.members?.length || 0) > 0
   );
   const builder = new xml2js.Builder();
   const updatedFileContent = builder.buildObject(manifest);
@@ -150,7 +191,7 @@ export async function catchMatches(
       }
       const catcherLabel = catcher.regex
         ? `regex ${catcher.regex.toString()}`
-        : 'ERROR';
+        : "ERROR";
       matchResults.push({
         fileName,
         fileText,
@@ -175,7 +216,7 @@ export async function countRegexMatches(
   regex: RegExp,
   text: string
 ): Promise<number> {
-  return ((text || '').match(regex) || []).length;
+  return ((text || "").match(regex) || []).length;
 }
 
 // Get all captured groups of a regex in a string
@@ -183,8 +224,8 @@ export async function extractRegexGroups(
   regex: RegExp,
   text: string
 ): Promise<string[]> {
-  const matches = ((text || '').match(regex) || []).map(e =>
-    e.replace(regex, '$1').trim()
+  const matches = ((text || "").match(regex) || []).map(e =>
+    e.replace(regex, "$1").trim()
   );
   return matches;
   // return ((text || '').matchAll(regex) || []).map(item => item.trim());
@@ -196,31 +237,28 @@ export async function generateReports(
   columns: any[],
   commandThis: any
 ): Promise<any[]> {
-  const logFileName = 'sfdx-hardis-' + commandThis.id.substr(commandThis.id.lastIndexOf(':') + 1);
-  const reportFile = path.resolve(
-    `./hardis-report/${logFileName}.csv`
-  );
-  const reportFileExcel = path.resolve(
-    `./hardis-report/${logFileName}.xls`
-  );
+  const logFileName =
+    "sfdx-hardis-" + commandThis.id.substr(commandThis.id.lastIndexOf(":") + 1);
+  const reportFile = path.resolve(`./hardis-report/${logFileName}.csv`);
+  const reportFileExcel = path.resolve(`./hardis-report/${logFileName}.xls`);
   await fs.ensureDir(path.dirname(reportFile));
   const csv = csvStringify(resultSorted, {
-    delimiter: ';',
+    delimiter: ";",
     header: true,
     columns
   });
-  await fs.writeFile(reportFile, csv, 'utf8');
+  await fs.writeFile(reportFile, csv, "utf8");
   const excel = csvStringify(resultSorted, {
-    delimiter: '\t',
+    delimiter: "\t",
     header: true,
     columns
   });
-  await fs.writeFile(reportFileExcel, excel, 'utf8');
-  commandThis.ux.log('[sfdx-hardis] Generated report files:');
+  await fs.writeFile(reportFileExcel, excel, "utf8");
+  commandThis.ux.log("[sfdx-hardis] Generated report files:");
   commandThis.ux.log(`[sfdx-hardis] - CSV: ${reportFile}`);
   commandThis.ux.log(`[sfdx-hardis] - XLS: ${reportFileExcel}`);
   return [
-    { type: 'csv', file: reportFile },
-    { type: 'xls', file: reportFileExcel }
+    { type: "csv", file: reportFile },
+    { type: "xls", file: reportFileExcel }
   ];
 }