+++ src/config/index.ts
@@ -1,14 +1,21 @@
-import * as child from 'child_process';
+/*
+sfdx-hardis is managed in 3 layers, wit hthe following priority
+- project, stored in /config
+- branches, stored in /config/branches
+- user, stored in /config/users
+
+getConfig(layer) returns:
+- project + branches + user if layer is user
+- project + branches if layer is branch
+- project if layer is project
+*/
+
 import { cosmiconfig } from 'cosmiconfig';
 import * as fs from 'fs-extra';
+import * as path from "path";
 import * as yaml from 'js-yaml';
 import * as os from 'os';
-import simpleGit, {SimpleGit} from 'simple-git';
-let git: SimpleGit = null ;
-const isGit = child.exec('git rev-parse --is-inside-work-tree 2>/dev/null', {encoding: 'utf8'});
-if (isGit) {
-    git = simpleGit();
-}
+import { getCurrentGitBranch, isGit } from '../common/utils';
 
 const moduleName = 'sfdx-hardis';
 const projectConfigFiles = [
@@ -20,36 +27,35 @@ const projectConfigFiles = [
 ];
 const username = os.userInfo().username;
 const userConfigFiles = [
-    `config/.${moduleName}.${username}.yaml`,
-    `config/.${moduleName}.${username}.yml`
+    `config/users/.${moduleName}.${username}.yaml`,
+    `config/users/.${moduleName}.${username}.yml`
 ];
 
 async function getBranchConfigFiles() {
     if (!isGit) {
         return [];
     }
-    const gitBranch = (await git.branchLocal()).current;
-    const gitBranchFormatted = gitBranch.replace('/', '-');
+    const gitBranchFormatted = getCurrentGitBranch({formatted:true});
     const branchConfigFiles = [
-        `config/.${moduleName}.${gitBranchFormatted}.yaml`,
-        `config/.${moduleName}.${gitBranchFormatted}.yml`
+        `config/branches/.${moduleName}.${gitBranchFormatted}.yaml`,
+        `config/branches/.${moduleName}.${gitBranchFormatted}.yml`
     ];
     return branchConfigFiles;
 }
 
-export const getConfig = async (layer: string = 'project'): Promise<any> => {
+export const getConfig = async (layer: string = 'user'): Promise<any> => {
     const defaultConfig = await loadFromConfigFile(projectConfigFiles);
     if (layer === 'project') {
         return defaultConfig;
     }
-    let userConfig = await loadFromConfigFile(userConfigFiles);
-    userConfig = Object.assign(defaultConfig, userConfig);
-    if (layer === 'user') {
-        return userConfig;
-    }
     let branchConfig = await loadFromConfigFile(await getBranchConfigFiles());
     branchConfig = Object.assign(defaultConfig, branchConfig);
-    return branchConfig;
+    if (layer === "branch") {
+        return branchConfig;
+    }
+    let userConfig = await loadFromConfigFile(userConfigFiles);
+    userConfig = Object.assign(defaultConfig, userConfig);
+    return userConfig;
 };
 
 // Set data in configuration file
@@ -80,5 +86,6 @@ async function setInConfigFile(searchPlaces: string[], propValues: any) {
         doc = yaml.load(fs.readFileSync(configFile, 'utf8'));
     }
     doc = Object.assign(doc, propValues);
+    await fs.ensureDir(path.dirname(configFile));
     await fs.writeFile(configFile, yaml.dump(doc));
 }